requirements.txt
pandas==2.1.0
pyspark==3.5.0
kafka-python==2.0.2
requests==2.31.0
pytest==7.4.0
azure-storage-file-datalake==12.14.0
azure-identity==1.14.0
python-dotenv==1.0.0
